Move Alert model to studies
