+++ src/main/java/org/sagebionetworks/bridge/models/studies/Alert.java
@@ -1,4 +1,4 @@
-package org.sagebionetworks.bridge.models.alerts;
+package org.sagebionetworks.bridge.models.studies;
 
 import javax.persistence.AttributeOverride;
 import javax.persistence.AttributeOverrides;
